extract positional_printer class to own file(positional_printer.hpp)
